moved ldap service stuff into dedicated folder
