+++ app/services/ldap/LdapUser.scala
@@ -1,0 +1,3 @@
+package services.ldap
+
+final case class LdapUser(systemId: String, lastname: String, firstname: String, email: String, status: String, registrationId: Option[String], degreeAbbrev: Option[String])
